+++ core/src/com/unciv/ui/worldscreen/mainmenu/OptionsPopup.kt
@@ -3,7 +3,6 @@ package com.unciv.ui.worldscreen.mainmenu
 import com.badlogic.gdx.Application
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Input
-import com.badlogic.gdx.Net
 import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.scenes.scene2d.ui.Label
 import com.badlogic.gdx.scenes.scene2d.ui.SelectBox
@@ -15,6 +14,7 @@ import com.unciv.MainMenuScreen
 import com.unciv.UncivGame
 import com.unciv.logic.MapSaver
 import com.unciv.logic.civilization.PlayerType
+import com.unciv.logic.multiplayer.SimpleHttp
 import com.unciv.models.UncivSound
 import com.unciv.models.metadata.BaseRuleset
 import com.unciv.models.ruleset.Ruleset
@@ -35,15 +35,7 @@ import com.unciv.ui.utils.*
 import com.unciv.ui.utils.LanguageTable.Companion.addLanguageTables
 import com.unciv.ui.utils.UncivTooltip.Companion.addTooltip
 import com.unciv.ui.worldscreen.WorldScreen
-import java.io.BufferedReader
-import java.io.DataOutputStream
-import java.io.InputStreamReader
-import java.net.DatagramSocket
-import java.net.HttpURLConnection
-import java.net.InetAddress
-import java.net.URL
-import java.nio.charset.Charset
-import java.util.*
+import java.util.UUID
 import kotlin.math.floor
 import com.badlogic.gdx.utils.Array as GdxArray
 
@@ -131,6 +123,10 @@ class OptionsPopup(val previousScreen: BaseScreen) : Popup(previousScreen) {
             previousScreen.game.setScreen(MainMenuScreen())
         }
         (previousScreen.game.screen as BaseScreen).openOptionsPopup()
+    }
+
+    private fun successfullyConnectedToServer(action: (Boolean, String)->Unit){
+        SimpleHttp.sendGetRequest("${settings.multiplayerServer}/isalive", action)
     }
 
     //region Page builders
@@ -295,52 +291,6 @@ class OptionsPopup(val previousScreen: BaseScreen) : Popup(previousScreen) {
                 popup.addCloseButton()
             }
         }).row()
-    }
-
-    fun getIpAddress(): String? {
-        DatagramSocket().use { socket ->
-            socket.connect(InetAddress.getByName("8.8.8.8"), 10002)
-            return socket.localAddress.hostAddress
-        }
-    }
-
-    object SimpleHttp {
-        fun sendGetRequest(url: String, action: (success: Boolean, result: String)->Unit) {
-            sendRequest(Net.HttpMethods.GET, url, "", action)
-        }
-
-        fun sendRequest(method: String, url: String, content: String, action: (success: Boolean, result: String)->Unit) {
-            with(URL(url).openConnection() as HttpURLConnection) {
-                requestMethod = method  // default is GET
-
-                if (method != Net.HttpMethods.GET) doOutput = true
-
-                try {
-                    if (content.isNotEmpty()) {
-                        // StandardCharsets.UTF_8 requires API 19
-                        val postData: ByteArray = content.toByteArray(Charset.forName("UTF-8"))
-                        val outputStream = DataOutputStream(outputStream)
-                        outputStream.write(postData)
-                        outputStream.flush()
-                    }
-
-                    val text = BufferedReader(InputStreamReader(inputStream)).readText()
-                    action(true, text)
-                } catch (t: Throwable) {
-                    println(t.message)
-                    val errorMessageToReturn =
-                        if (errorStream != null) BufferedReader(InputStreamReader(errorStream)).readText()
-                        else t.message!!
-                    println(errorMessageToReturn)
-                    action(false, errorMessageToReturn)
-                }
-            }
-        }
-
-    }
-    
-    fun successfullyConnectedToServer(action: (Boolean, String)->Unit){
-        SimpleHttp.sendGetRequest("${settings.multiplayerServer}/isalive", action)
     }
 
     private fun getAdvancedTab() = Table(BaseScreen.skin).apply {
